extracted remaining class to generate the meta classes
